🐛 Fixing connections not showing up in dev environment
