+++ packages/editor-ui/src/components/Node/NodeCreation.vue
@@ -18,7 +18,7 @@
 
 <script lang="ts">
 import Vue from "vue";
-import { getMidCanvasPosition } from '@/utils';
+import { getMidCanvasPosition } from '@/utils/nodeViewUtils';
 import {DEFAULT_STICKY_HEIGHT, DEFAULT_STICKY_WIDTH, STICKY_NODE_TYPE} from "@/constants";
 import { mapStores } from "pinia";
 import { useUIStore } from "@/stores/ui";
